requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
Pillow==11.0.0
anthropic==0.21.3
httpx==0.24.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.2.1
opencv-python-headless==4.10.0.84
numpy==2.0.2
email-validator==2.1.0
jinja2==3.1.4

# Force rebuild 10/27/2025 07:01:16
